pyproject.toml
[project]
name = "iran-sentiment"
version = "0.1.0"
description = "Sentiment analysis experiment: US administration messaging around the 2026 Iran war"
requires-python = ">=3.11"
dependencies = [
    "tweepy>=4.14,<5",            # X / Twitter API v2
    "httpx>=0.27",                 # async HTTP for Truth Social scraping
    "transformers>=4.40",          # HuggingFace sentiment models
    "torch>=2.2",                  # backend for transformers
    "vaderSentiment>=3.3",         # rule-based sentiment (fast baseline)
    "pandas>=2.2",
    "matplotlib>=3.9",
    "seaborn>=0.13",
    "plotly>=5.20",
    "python-dotenv>=1.0",
    "tqdm>=4.66",
    "click>=8.1",
]

[project.optional-dependencies]
dev = ["pytest", "ruff", "ipykernel", "jupyter"]
truthsocial = ["truthbrush>=0.2,<0.3"]  # Stanford IO Truth Social client

[project.scripts]
iran-sentiment = "src.cli:main"

[tool.ruff]
line-length = 100
target-version = "py311"
